Fix compilation of nym-vpn-lib-types, move vpnd types to nym-vpnd-types
