+++ nym-vpn-core/crates/nym-vpn-lib-types/src/lib.rs
@@ -5,7 +5,6 @@
 
 mod account;
 mod connection_data;
-mod service;
 mod tunnel_event;
 mod tunnel_state;
 
@@ -26,7 +25,6 @@ pub use connection_data::{
     ConnectionData, Gateway, MixnetConnectionData, NymAddress, TunnelConnectionData,
     WireguardConnectionData, WireguardNode,
 };
-pub use service::{VpnServiceConnectError, VpnServiceDisconnectError, VpnServiceInfo};
 pub use tunnel_event::{
     BandwidthEvent, ConnectionEvent, ConnectionStatisticsEvent, MixnetEvent, SphinxPacketRates,
     TunnelEvent,
