+++ nym-vpn-core/crates/nym-vpnd-types/src/lib.rs
@@ -2,3 +2,4 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 pub mod gateway;
+pub mod service;
